requirements.txt
aiogram
requests
python-telegram-bot-pagination